+++ src/main/java/racingcar/di/CarFactoryImpl.java
@@ -1,5 +1,9 @@
-package racingcar;
+package racingcar.di;
 
+import racingcar.domain.Car;
+import racingcar.domain.Engine;
+import racingcar.domain.NormalEngine;
+import racingcar.domain.RacingCar;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -8,8 +12,9 @@ public class CarFactoryImpl implements CarFactory{
     @Override
     public List<Car> makeCars(String[] names) {
         List<Car> cars = new ArrayList<>();
+        Engine normalEngine = new NormalEngine();
         for (String name : names) {
-            Car car = new RacingCar(name);
+            Car car = new RacingCar(name, normalEngine);
             cars.add(car);
         }
         return cars;
